Migrar funções e arquivo config
